pyproject.toml
[project]
name = "my-mirror-on-cloud"
version = "0.1.0"
description = "Short project summary"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "numpy (>=2.3.2,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "ollama (>=0.5.3,<0.6.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "weaviate-client (>=4.16.9,<5.0.0)",
    "fashion-clip (>=0.2.2,<0.3.0)",
    "loguru (>=0.7.3,<0.8.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
ruff = "^0.12.11"
pre-commit = "^4.3.0"
mkdocs = "^1.6.1"
ipykernel = "^6.30.1"
nbstripout = "^0.8.1"
codespell = "^2.4.1"
ipywidgets = "^8.1.7"
widgetsnbextension = "^4.0.14"
matplotlib = "^3.10.6"
seaborn = "^0.13.2"
plotly-express = "^0.4.1"
dotenv = "^0.9.9"
